chore: move partners listing into a separate file
